refactor: Split out overlap checking in Schedule
